feat(rpi4): Wakeonlan systemd service
In case of a long black out where the UPS' battery rans out, the main
server will power off. When the power comes back up, I would like the
server to turn back on.

To do that, I have my Raspberry Pi send a wake-on-lan packet after
the network is up.

My limited testing shows that this works fine.

If it turns out `systemd`'s `network-online.target` doesn't work, it
might be better to use a `systemd` timer along side the service that
will trigger 5-10 minutes after the RPi has booted. That should be
plenty of time for the network to be up and for the server to turn on.