fix(nix): Use a tmpfs for /tmp, but build in /var/tmp
Having `/tmp` on a tmpfs is the normal behavior a user expects. The
issue with Nix is that it uses `$TMPDIR` to store build "artifacts".
With `/tmp` being a tmpfs, its size is limited by the amount of RAM
the host has, often resulting in an "device out of space" error when
building stuff. To solve this, we configure `nix-daemon` to store
build artifacts in `/var/tmp`.