feat: Back to fish
I think I overall prefer zsh, but I need to figure out a reliable way
of generating completions for packages installed in dev shells.
In fish, I can just generate the completion and put it in
`~/.confign/fish/completions,` but in zsh, you have to modify `fpath`,
which apparently cannot be done in a `shellHook` (since the shell
used in the dev shell is `bash`).